CON-47: Custom serialization to Protobuf

+++ server/src/main/kotlin/com/fyordo/cms/server/config/PropertyVersionConfig.kt
@@ -5,6 +5,6 @@ import org.springframework.context.annotation.Configuration
 
 @ConfigurationProperties(prefix = "app.property")
 data class PropertyVersionConfig(
-    val currentVersion: Byte = 1,
+    val currentVersion: Int = 1,
     val maxValueSizeBytes: Int = 1024 * 1024 // 1MB default
 )
